+++ examples/showoci/CHANGELOG.rst
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 =====================
+21.06.29 - 21.06.29
+=====================
+* Added DRG Route Tables for DRGV2
+* Added Additional database attributes
+
+=====================
 21.06.02 - 21.06.02
 =====================
 * Support for New DRG